c 0.8

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
  * Return: coun.
  */
 
-int print_list(const list_t *h)
+size_t print_list(const list_t *h)
 {
 	int coun = 0;
