+++ 0x12-singly_linked_lists/lists.h
@@ -11,6 +11,6 @@ typedef struct lists
 
 
 int _putchar(char c);
-int print_list(const list_t *h);
+size_t print_list(const list_t *h);
 
 #endif
